lr_scheduler

+++ abyss/training/metrics.py
@@ -4,9 +4,9 @@ import torchmetrics
 
 def metric_accuracy(output: torch.Tensor, label: torch.Tensor) -> torch.Tensor:
     """Log accuracy for certain stage"""
-    return torchmetrics.functional.classification.accuracy(output.to(torch.float32), label.to(torch.int8))
+    return torchmetrics.functional.classification.accuracy(output, label)
 
 
 def metric_dice(output: torch.Tensor, label: torch.Tensor) -> torch.Tensor:
     """Log dice for certain stage"""
-    return torchmetrics.functional.classification.dice(output.to(torch.float32), label.to(torch.int8))
+    return torchmetrics.functional.classification.dice(output.to(torch.int8), label.to(torch.int8))
